Algorithms

Algorithms

The first argument of an Algorithm's constructor is an SModel. This is to ensure storage buffers are the correct size.

SparseRegression.ProxGrad — Type.

ProxGrad(model, step = 1.0)

Proximal gradient method with step size step. Works for any loss and any penalty with a prox method.

source

SparseRegression.Fista — Type.

Fista(model, step = 1.0)

Accelerated proximal gradient method. Works for any loss and any penalty with a prox method.

source

SparseRegression.GradientDescent — Type.

GradientDescent(model, step = 1.0)

Gradient Descent. Works for any loss and any penalty.

source

SparseRegression.Sweep — Type.

Sweep(model)

Linear/ridge regression via sweep operator. Works for (scaled) L2DistLoss with NoPenalty or L2Penalty.

source

SparseRegression.LinRegCholesky — Type.

LinRegCholesky(model)

Linear/ridge regression via cholesky decomposition. Works for (scaled) L2DistLoss with NoPenalty or L2Penalty.

source